+++ packages/pangea-sdk/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Sanitize now defaults to the post-url transfer method.
+
 ## 3.8.0-beta.4 - 2024-03-28
 
 ### Fixed